chore: use goerli instead of gnosis chain
